bux/requirements.txt
asgiref==3.1.2
async-timeout==3.0.1
attrs==19.1.0
autobahn==19.3.3
Automat==0.7.0
certifi==2017.11.5
channels==2.2.0
chardet==3.0.4
constantly==15.1.0
daphne==2.3.0
Django==2.2.1
enum34==1.1.6
Faker==1.0.6
hyperlink==19.0.0
idna==2.6
incremental==17.5.0
nba-data==1.9
PyExecJS==1.5.1
PyHamcrest==1.9.0
pypiwin32==223
PyReact==0.6.0
python-dateutil==2.8.0
pytz==2019.1
pywin32==224
PyYAML==5.1
requests==2.21.0
six==1.12.0
sqlparse==0.3.0
text-unidecode==1.2
Twisted==19.2.0
txaio==18.8.1
ujson==1.35
urllib3==1.24.3
v20==3.0.25.0
virtualenv==16.5.0
zope.interface==4.6.0
